requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
